Changing int to long long int

+++ A1/q1/q1.c
@@ -2,16 +2,16 @@
 #include <stdlib.h>
 
 
-extern int find_lonely(int* arr,int n);
+extern long long int find_lonely(long long int* arr,int n);
 
 
 int main(){
     int n;
     scanf("%d",&n);
     n = 2*n + 1;
-    int* arr = (int*)malloc(n*sizeof(int));
+    long long int* arr = (long long int*)malloc(n*sizeof(long long int));
     for(int i =0; i<n; i++){
-        scanf("%d",&arr[i]);
+        scanf("%lld",&arr[i]);
     }
-    printf("%d",find_lonely(arr,n));
+    printf("%lld",find_lonely(arr,n));
 }
